requirements.txt
fastapi>=0.109
uvicorn>=0.27
sqlalchemy>=2.0
aiosqlite>=0.20
pydantic>=2.0
pydantic-settings>=2.0
httpx>=0.27
pytest>=8.0
pytest-asyncio>=0.23,<1.0
greenlet>=3.0
pyyaml>=6.0
